refactor(coupon) : 패키지 수정
